python/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
fastapi==0.116.2
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-billing==1.16.3
google-cloud-common==1.5.2
google-cloud-compute==1.37.0
google-cloud-core==2.4.3
google-cloud-filestore==1.13.2
google-cloud-lustre==0.1.2
google-cloud-netapp==0.3.24
google-cloud-parallelstore==0.2.15
google-cloud-quotas==0.1.18
google-cloud-storage==3.4.0
google-cloud-tpu==1.23.2
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.75.0
grpcio-status==1.75.0
h11==0.16.0
idna==3.10
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
